refactor(controlplane): move implicit sample configuration initialization
